+++ source/runner/templates/_tests/_bitbucketCloudTemplate.test.ts
@@ -10,6 +10,7 @@ import {
 } from "../../_tests/fixtures/ExampleDangerResults"
 import { dangerSignaturePostfix, template, inlineTemplate } from "../bitbucketCloudTemplate"
 import { DangerResults } from "../../../dsl/DangerResults"
+import * as utils from "../../DangerUtils"
 
 const commitID = "e70f3d6468f61a4bef68c9e6eaba9166b096e23c"
 
@@ -18,11 +19,22 @@ const warningEmoji = "⚠️"
 const messageEmoji = "✨"
 
 describe("generating messages for BitBucket cloud", () => {
+  let complimentMock: jest.Mock<string, []>
+  beforeEach(() => {
+    complimentMock = jest.spyOn(utils, "compliment").mockReturnValue("Well done.")
+  })
+
+  afterEach(() => {
+    complimentMock.mockRestore()
+  })
+
   it("shows no sections for empty results", () => {
     const issues = template("blankID", emptyResults, commitID)
     expect(issues).not.toContain("Fails")
     expect(issues).not.toContain("Warnings")
     expect(issues).not.toContain("Messages")
+    expect(issues).toContain("Well done.")
+    expect(complimentMock).toBeCalledTimes(1)
   })
 
   it("shows no sections for results without messages", () => {
@@ -30,18 +42,24 @@ describe("generating messages for BitBucket cloud", () => {
     expect(issues).not.toContain("Fails")
     expect(issues).not.toContain("Warnings")
     expect(issues).not.toContain("Messages")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
   })
 
   it("Shows the failing messages in a section", () => {
     const issues = template("blankID", failsResults, commitID)
     expect(issues).toContain("Fails")
     expect(issues).not.toContain("Warnings")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
   })
 
   it("Shows the warning messages in a section", () => {
     const issues = template("blankID", warnResults, commitID)
     expect(issues).toContain("Warnings")
     expect(issues).not.toContain("Fails")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
   })
 
   it("summary result matches snapshot, with a commit", () => {
@@ -62,6 +80,10 @@ describe("generating messages for BitBucket cloud", () => {
 
   it("shows a postfix message with no commit ID if not provided", () => {
     expect(template("blankID", emptyResults)).toContain(dangerSignaturePostfix({} as DangerResults))
+  })
+
+  it("empty result matches snapshot, with a compliment", () => {
+    expect(template("blankID", emptyResults, commitID)).toMatchSnapshot()
   })
 })